+++ src/main/java/org/delia/db/newhls/ValType.java
@@ -1,0 +1,16 @@
+package org.delia.db.newhls;
+
+/**
+ * Values in queries. note dates are strings in delia.
+ * @author ian
+ *
+ */
+public enum ValType {
+	BOOLEAN,
+	INT,
+	LONG,
+	NUMBER,
+	STRING,
+	SYMBOL,
+	FUNCTION
+}
