+++ src/main/java/org/delia/db/newhls/cond/FilterCond.java
@@ -1,0 +1,8 @@
+package org.delia.db.newhls.cond;
+
+/**
+ * A filter condition. eg what goes inside Customer[]
+ */
+public interface FilterCond {
+
+}
