+++ src/main/java/org/delia/db/newhls/cond/FilterOp.java
@@ -1,0 +1,19 @@
+package org.delia.db.newhls.cond;
+
+/**
+ * One of the six operations. eg size > 10
+ * @author ian
+ *
+ */
+public class FilterOp {
+	public String op; //==,!=,<,>,<=,>=
+
+	public FilterOp(String op) {
+		this.op = op;
+	}
+
+	@Override
+	public String toString() {
+		return op;
+	}
+}
